rename useLaunchpadChecklist file name to use-launchpad-checklist.tsx
