+++ packages/help-center/src/hooks/use-launchpad-checklist.tsx
